requirements.txt
cycler==0.10.0
future==0.18.2
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.3
Pillow==7.1.1
pkg-resources==0.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.14.0
torch==1.5.0
torchvision==0.6.0
tqdm==4.45.0
